GH-1260: Add exception classification to DARP
Resolves https://github.com/spring-projects/spring-kafka/issues/1260

Refactor common code with `SeekToCurrentErrorHandler` into
`FailedRecordProcessor` super class.
